Ticket 576: Memory usage constantly grows when running fm12_importer
The cause for the growing memory usage was that sqlalchemys sesionmaker was created several times for each handled file. According to sqlalchemy:s documentation, it should only be created once per application. Code has with this fix been updated to only create one sessionmaker and reuse that throughout the application.