Ticket 577: Kill command to running FM12 daemon has no effect
Kill command had no effect since the script first checked whether the importer was already alive, and if so issued a warning and exited. The code for handling the kill-command was thus never reached if the importer was running. With this change, the order of the checks is reversed. Kill-command handling is done before it is checked whether the importer is alive or not.