Ticket 646: Error in gra-adjustment of composites
Gain value of distance field in ACRR-products changed from 1.0 to 1000.0. This gives a converted value in unit meters instead of kilometers. The GRA-adjustment algorithm also updated to expect values in unit meters.

The erroneous values of the gra-adjusted field in the composites was caused by the fact that the distance field multiplied with the distance gain in the composites resulted in a value in unit meters, while the GRA-adjustment algorithm expected values in km as input. With the above changes, meters is used throughout the code.