Ticket 650: Quality field attributes
Quality fields in composites now include gain and offset attributes, giving them a correct range between 0 and 1, if converted by applying gain- and offset-values.

When writing the quality fields in the composite, the values in the fields are now converted to all use the same predefined gain and offset values. The gain value is set to 1/255 and offset is set to 0. This means that the values in the data set will have values between 0 and 255, thus using the full range of the 8-bit integer field they are stored in.