Ticket 672: rave-overshooting crash issue
The root cause for the crash in rave_overshooting_quality_plugin.py was that it was provided a volume without any scans in it. The empty volumes occured if all scans of it were marked as 'malfunc' and 'ignore malfunc' was enabled in the compositing rule. In those cases, all scans were removed from the volume, but the volume remained. The implemented solution removes the complete volume in cases where all scans of the volume are 'malfunc'. This avoids the crash in the rave_overshooting-plugin.

In addition to fixing the above problem, this commit also removes exceptions thrown for site2d-rules, in cases where the all relevant scans are malfunc. Furthermore, related problems are fixed for tiled composite generation. More specifically, problems occurred if e.g., all scans/volumes of one tile were removed due to 'malfunc'-marking. These are solved with these commit.