Ticket 674: Add support for removing malfunc scans in ApplyQc-plugin
Added argument --remove-malfunc to rave_pgf_apply_qc_plugin. If this is set to true, all scans that are marked as 'malfunc' will be removed from the volume prior to applying the QCs to it. If all scans of the volume are marked as 'malfunc', no new volume will be generated. The same is true if the volume itself is marked 'malfunc'.