Ticket 727: Volume with unsorted elangles gives faulty CAPPI/PCAPPI composites
Compositing algorithm updated to always perform a sort on the scans of an input volume. The sort is an ascending sort based on the elangle values of the scans. This ensures that CAPPI and PCAPPI products will be correctly generated. Data from the correct scans will be used, even if the input volume has not sorted scans.